pyproject.toml
[project]
name = "api-nasa"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
